requirements.txt
lxml
docopt
tqdm >= 2.0
tldextract
requests
sklearn-crfsuite >= 0.3.1

# 0.17 is need for LabelKFold
scikit-learn >= 0.17

# "formtype/tool.py evaluate" command uses pandas for pretty printing;
# if you don't plan to contribute to formtype library pandas is not needed.
pandas
